Ajustements

+++ medical/gql_mutations.py
@@ -112,8 +112,8 @@ def reset_item_or_service_before_update(item_service):
 
 
 def update_or_create_item_or_service(data, user, item_service_model):
-    items = data.pop('items') if 'items' in data else []
-    services = data.pop('services') if 'services' in data else []
+    items = data.pop('items') if 'items' in data else None
+    services = data.pop('services') if 'services' in data else None
     client_mutation_id = data.pop('client_mutation_id', None)
     data.pop('client_mutation_label', None)
     item_service_uuid = data.pop('uuid') if 'uuid' in data else None
@@ -129,37 +129,39 @@ def update_or_create_item_or_service(data, user, item_service_model):
 
     if item_service_uuid:
         item_service = item_service_model.objects.get(uuid=item_service_uuid)
-        # Delete Service present in the Database and absent in the list sent by FE
-        # Means that user click on delete button and old Service is not sent
-        serviceExisting = list()
-        serviceSent = list()
-        for ServiceList in ServiceService.objects.filter(servicelinkedService=item_service.id).all() :
-            serviceExisting.append(ServiceList.id)
+        if items:
+            # Delete Service present in the Database and absent in the list sent by FE
+            # Means that user click on delete button and old Service is not sent
+            serviceExisting = list()
+            serviceSent = list()
+            for ServiceList in ServiceService.objects.filter(servicelinkedService=item_service.id).all() :
+                serviceExisting.append(ServiceList.id)
 
-        for ServiceList in services:
-            serviceSent.append(ServiceList.id)
+            for ServiceList in services:
+                serviceSent.append(ServiceList.id)
 
-        serviceToDelete = list(set(serviceExisting) - set(serviceSent))
-        for serviceToDeleteId in serviceToDelete:
-            ServiceService.objects.filter(
-                id=serviceToDeleteId,
-            ).delete()
+            serviceToDelete = list(set(serviceExisting) - set(serviceSent))
+            for serviceToDeleteId in serviceToDelete:
+                ServiceService.objects.filter(
+                    id=serviceToDeleteId,
+                ).delete()
 
-        # Delete Item present in the Database and absent in the list sent by FE
-        # Means that user click on delete button and old Ites is not sent
-        itemExisting = list()
-        itemSent = list()
-        for ItemList in ServiceItem.objects.filter(servicelinkedItem=item_service.id).all() :
-            itemExisting.append(ItemList.id)
+        if services:
+            # Delete Item present in the Database and absent in the list sent by FE
+            # Means that user click on delete button and old Ites is not sent
+            itemExisting = list()
+            itemSent = list()
+            for ItemList in ServiceItem.objects.filter(servicelinkedItem=item_service.id).all() :
+                itemExisting.append(ItemList.id)
 
-        for ItemList in items:
-            itemSent.append(ItemList.id)
+            for ItemList in items:
+                itemSent.append(ItemList.id)
 
-        itemToDelete = list(set(itemExisting) - set(itemSent))
-        for itemToDeleteId in itemToDelete:
-            ServiceItem.objects.filter(
-                id=itemToDeleteId,
-            ).delete()
+            itemToDelete = list(set(itemExisting) - set(itemSent))
+            for itemToDeleteId in itemToDelete:
+                ServiceItem.objects.filter(
+                    id=itemToDeleteId,
+                ).delete()
         reset_item_or_service_before_update(item_service)
         [setattr(item_service, key, data[key]) for key in data]
     else:
@@ -211,8 +213,8 @@ class CreateOrUpdateItemOrServiceMutation(OpenIMISMutation):
         data['audit_user_id'] = user.id_for_audit
         from core.utils import TimeUtils
         data['validity_from'] = TimeUtils.now()
-        print("Create or Update Item or Service Mutation");
-        print(data);
+        print("Create or Update Item or Service Mutation")
+        print(data)
         update_or_create_item_or_service(data, user, cls.item_service_model)
         return None
